Sales report feature added; security of app updated

+++ app/controllers/orders_controller.rb
@@ -1,16 +1,13 @@
 class OrdersController < ApplicationController
   before_action :set_order, only: [:show, :edit, :update, :destroy]
+  before_action :ensure_owner_or_clerk, only: [:index, :update, :edit]
+  before_action :ensure_owner, only: [:report]
 
   # GET /orders
   # GET /orders.json
   def index
     @orders = Order.pending_order.order(date: :desc)
     @your_order = false
-  end
-
-  # GET /orders/1
-  # GET /orders/1.json
-  def show
   end
 
   def your_orders
@@ -20,10 +17,30 @@ class OrdersController < ApplicationController
   end
 
   def report
-    @orders = Order.order(date: :desc)
-    @totals = OrderItem.group(:order_id).sum(:total)
-    @count = Order.count
-    @sum = OrderItem.sum(:total)
+    if (params[:to_date])
+      @orders = []
+      if (params[:to_date] == "" or params[:from_date] == "")
+        flash[:error] = "Date can't be empty"
+        redirect_to sales_report_path
+      elsif (params[:from_date] > params[:to_date])
+        flash[:error] = "Invalid Date range"
+        redirect_to sales_report_path
+      else
+        @from_date = params[:from_date].split("-").join("/")
+        @from_date = Date.strptime(@from_date, "%Y/%m/%d")
+        @to_date = params[:to_date].split("-").join("/")
+        @to_date = Date.strptime(@to_date, "%Y/%m/%d")
+        @orders = Order.get_report(params[:from_date], params[:to_date], params[:user_id])
+        @totals = OrderItem.group(:order_id).sum(:total)
+        @count = @orders.count
+        @sum = @orders.map { |item| item.order_items.sum(:total) }.sum
+      end
+    else
+      @orders = Order.all
+      @totals = OrderItem.group(:order_id).sum(:total)
+      @count = Order.count
+      @sum = OrderItem.sum(:total)
+    end
   end
 
   # GET /orders/new
@@ -47,12 +64,15 @@ class OrdersController < ApplicationController
   # POST /orders.json
   def create
     items = @current_user.carts
-    if items.count != 0
+    type = "walk in"
+    type = "Online" if @current_user.is_customer(@current_user.id)
+    if items.count > 0
       new_order = Order.create!(
         user_id: @current_user.id,
         date: Time.now.getutc,
         delivered_at: nil,
         status: "confirm",
+        order_type: type,
       )
     end
 
@@ -66,8 +86,13 @@ class OrdersController < ApplicationController
         total: item.quantity * item.menu_item.menu_item_price,
       )
     end
-    items.destroy_all
-    redirect_to yourorder_path
+    if items.count > 0
+      items.destroy_all
+      redirect_to yourorder_path(notice: "Order Successfully placed")
+    else
+      flash[:error] = "Your Cart is empyt !!"
+      redirect_to carts_path
+    end
   end
 
   # PATCH/PUT /orders/1
@@ -82,16 +107,6 @@ class OrdersController < ApplicationController
     end
 
     redirect_to orders_path(notice: "Order Status Updated Successfully ")
-  end
-
-  # DELETE /orders/1
-  # DELETE /orders/1.json
-  def destroy
-    @order.destroy
-    respond_to do |format|
-      format.html { redirect_to orders_url, notice: "Order was successfully destroyed." }
-      format.json { head :no_content }
-    end
   end
 
   private
